package

+++ src/dataforge/mainConfig.py
@@ -2,8 +2,7 @@ import argparse
 import os
 import shutil
 import sys
-
-from pg import Pg
+from .pg import Pg
 
 
 class MainConfig:
@@ -14,14 +13,14 @@ class MainConfig:
     # parse command line args
     _parser = argparse.ArgumentParser()
     _parser.add_argument('--mode', default='import', choices=['import', 'init', 'query'], type=str,
-                        help='Operation mode: import, init, query')
-    _parser.add_argument('--source', type=str, help='Source folder', required=True, metavar='<Source Path>')
+                         help='Operation mode: import, init, query')
+    _parser.add_argument('--source', type=str, help='Source folder', metavar='<Source Path>')
 
     _parser.add_argument('--output', type=str, help='Output folder', metavar='<Output Path>')
-    _parser.add_argument('--pg_conn', type=str, help='Postgres connection string', metavar='<Postgres connection string>')
+    _parser.add_argument('--pg_conn', type=str, help='Postgres connection string',
+                         metavar='<Postgres connection string>')
 
     args = _parser.parse_args()
-
     if args.mode == "init":
         if args.pg_conn is None:
             print("--pg_conn parameter is required for --mode init")
@@ -29,6 +28,9 @@ class MainConfig:
         pg = Pg(args.pg_conn, initialize=True)
     else:
         pg = Pg()
+    if args.source is None:
+        print("--source parameter is not provided")
+        sys.exit(0)
     output_path = args.output if args.output else args.source + '_out'
     log_path = os.path.join(output_path, 'log.txt')
     shutil.rmtree(output_path, ignore_errors=True)
@@ -37,5 +39,3 @@ class MainConfig:
     os.makedirs(output_source_path)
     output_output_path = os.path.join(output_path, 'outputs')
     os.makedirs(output_output_path)
-
-    pg = Pg()  # connect to postgres
